refactored: addrbook_interface
